@@ -14,18 +14,49 @@
         }
     ],
     "require": {
-        "keboola/php-utils": "^2.1",
-        "php": ">=7.0"
+        "php": ">=7.4",
+        "ext-json": "*",
+        "keboola/php-utils": "^4.1"
     },
     "require-dev": {
-        "phpunit/phpunit": "^6.2",
-        "codeclimate/php-test-reporter": "^0.4",
-        "squizlabs/php_codesniffer": "^3.0"
+        "keboola/coding-standard": ">=9.0",
+        "php-parallel-lint/php-parallel-lint": "^1.2",
+        "phpstan/phpstan": "^0.12",
+        "phpunit/phpunit": "^9.5"
     },
     "autoload": {
-        "psr-0": {
-            "Keboola\\Code": "src/",
-            "Keboola\\Code\\Tests": "tests/"
+        "psr-4": {
+            "Keboola\\Code\\": "src"
         }
+    },
+    "autoload-dev": {
+        "psr-4": {
+            "Keboola\\Code\\Tests\\": "tests/phpunit"
+        }
+    },
+    "scripts": {
+        "tests-phpunit": "phpunit",
+        "tests": [
+            "@tests-phpunit"
+        ],
+
+        "phpstan": "phpstan analyse ./src ./tests --level=max --no-progress -c phpstan.neon",
+        "phpcs": "phpcs -n --ignore=vendor --extensions=php .",
+        "phpcbf": "phpcbf -n --ignore=vendor --extensions=php .",
+        "phplint": "parallel-lint -j 10 --exclude vendor .",
+        "build": [
+            "@phplint",
+            "@phpcs",
+            "@phpstan",
+            "@tests"
+        ],
+        "ci": [
+            "@composer validate --no-check-publish --no-check-all",
+            "@build"
+        ]
+    },
+    "config": {
+        "sort-packages": true,
+        "optimize-autoloader": true
     }
 }
